requirements.txt
Django
psycopg2-binary
django-environ
djangorestframework
django-filter
celery
redis
django-redis
factory-boy
Faker
gunicorn
django_cleanup
drf-spectacular
django-imagekit
djangorestframework-simplejwt
python-slugify
whitenoise[brotli]
django-jazzmin
django-ckeditor

#pip install -r requirements.txt -r requirements-dev.txt